Better computation of temperature sensor coefficients
Implement a better algorithm for computation of temperature
sensor coefficients using Steinhart–Hart equations. This will
provide more accurate resistance calculation based on the
simulated heater temperature.

+++ src/core/objects/thermistor.h
@@ -20,7 +20,7 @@
 #include <stdint.h>
 
 typedef struct {
-    double resistance;
+    float resistance;
     uint16_t adc;
     char pin[8];
 } thermistor_status_t;
